move audio loudness to AvatarData, change AvatarData to ScriptOwnership
